pyproject.toml
[project]
name = "dd-creator"
version = "0.1.0"
description = "A Gradio-based tool to help users create high-quality datasets for training diffusion models."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=6.0.0",
    "torch",
    "torchvision",
    "transformers>=4.48.0",
    "pillow",
    "numpy",
    "einops",
    "timm<=1.0.9",
    "scipy",
    "accelerate",
    "bitsandbytes",
    "gradio-client>=1.6.0",
]
